requirements.txt
colorama
pyusb
libusb
